created read_json_status method and did some adjustments
created the read_json_status method which loops through the data folder, catches all .json files which contains "status" in there names, read them and puts the websitename, onlinestatus and latency in an dictonary and returns this at the end of the method. In the future the bot will work with this dictonary to give back a statusupdate about the sites. It's extremly neccesarry that the syntax of the .json files the the data folder continues like it is. Otherwise the method would probably not work. Of course i added the read_json_status method to the __init__.py file to use it for the Bot.
Additionally i changed the name of Utils.py to utils.py and the name of Website.py to website_class.py just to have a uniformly syntax with the filenames. Finally I worked around with the "sys.path.append()"-path at the beginning of the Bot, so the Bot should run now on every OS.