status command update and added clear command
the !website_status command looks much better now because its embed as a seperate message and isn't any longer a simple message. Also i added a  "clear"-command which allows you to clear a whole channel. in the future the command will only be accessible with a specified permission/role that not any user will be able to clear the messages of a whole channel.
The last thing I now have to do, is to add a function/command that gives you a choosen role based on your study, so everyone on the server knows if you are a "Wirtschaftsinformatiker", "Informatiker" etc...