added help command and welcome message to new users, adjust read_json_data_bot function and online_offline_unstable
added a help command with all functions and infos about the bot for every user available (currently only filled with the iu_websites_status command ^^), wrote a first welcome message how it could look like when a new member will join the discord channel and what are his options on it. also i adjust the read_json_data_bot function, so it now will read the data from the normal json files named by the website because the [websitename]-data.json files don't exits anylonger. additional it reads now the last_offline_time of the website too and gives it back as an information to the user. don't know if the format of the time will be shown correct to the user because at the moment the website is only available so it just shows -1 as signal(bot transforms -1 to - for the user) .