added commands for all roles / adjust is_user_roles
added commands to all roles from the "big" server, they probably need to adjust when the server is online because i currently don't know the real case of the letters from the roles. additionally i adjust the "is_user_role"- function with the condition to check if the current amount of roles from a specific user is higher or equal then 2, to avoid multiple courses on one user. If a user enters a wrong role he have to contact an admin to change it manually.