multiple options to set a role, started editing the "help" command
added multiple options to avoid as good as possible missspelling when someone have to set a role. of course is it possible to add more options. i wasn't able to find a good solutions with "buttons" in discord that it would look nice and also work as it should. last step would be to adjust the "help" command in a understandable and good looking design for all users and add the right spelled rolenames for the "real"- server